--- a/ggml/src/ggml-vulkan/ggml-vulkan.cpp
+++ b/ggml/src/ggml-vulkan/ggml-vulkan.cpp
@@ -1741,8 +1741,9 @@ static void ggml_vk_load_shaders(vk_device& device) {
         s_warptile_mmq_int = { subgroup_size_32, 32, 32, 32, 32,       32, 2, 2, 1, 1, subgroup_size_8 };
 
         // chip specific tuning
-        if (device->architecture == AMD_GCN)
-            m_warptile_mmq = { 256, 64, 64, 32, 16, 16, 2, 2, tn_m, tk_m, 16 };
+        if (device->architecture == AMD_GCN) {
+            m_warptile_mmq = m_warptile_mmq_int = { 256, 64, 64, 32, 16, 16, 2, 2, 2, 1, 16 };
+        }
 
         l_mmq_wg_denoms = l_wg_denoms = {128, 128, 1 };
         m_mmq_wg_denoms = m_wg_denoms = { 64,  64, 1 };