@netrunnereve 2025-04-19
Okay I think is ready. The 16x16 tiles I'm using now perform like the 64x16 ones at a fixed clock speed but once I turn frequency scaling back it manages to clock higher than before and I get a 4% improvement in pp512 speed. Maybe the smaller tile sizes make it run more efficiently?

Since all the threads in the workgroup do the same calculations we don't necessarily have to make the shader's warp size match the subgroup size.