@netrunnereve 2025-04-20
> The most obvious difference can be seen in the test with test-backend-ops, where the FLOPS values drop by about 8 percent when n=512, except for f16 and f32.
Even at the application level, the generation time increases by about 4 percent when using stable-diffusion-v1.5 in stable-diffusion.cpp.

My first thought was that your integrated graphics only had two cores as that wouldn't handle 256 threads well, but that's not the case as you have eight cores. It's also not a Vega or FP16 issue as @0cc4m's card is doing fine.

Since your 5700U is a 15W chip this might actually be a power issue though. For example when prompt processing Llama 2 7B Q4_0 on master my 470 runs at 1.15 GHz and gets 171 t/s. With my PR it only runs at 1 GHz but gets 189 t/s, and both times I'm hitting the 130A TDC limit on my card. Please run with a real model and compare the GPU clock speeds and power levels with master. On Linux you can use radeontop and sensors for this, for Windows I have no idea :woman_shrugging:.

> Maybe, like Java and other VMs, we should store performance profile data for each device and apply it automatically?

Obviously that's the best option but it's a lot of work. Right now everyone can just submit tunes as PRs as it's not that hard to do.