@netrunnereve 2025-04-21
> The conclusion is that test-backend-ops.exe was about 16% slower on average at Q8_0,n=512, while sd.exe was only about 2% slower on average under real conditions.

Again as mentioned can you run a `llama-bench` with a real model instead of a `test-backend-ops`? I don't plan on looking into the stable diffusion results as that's basically a fork with an older version of GGML and possibly some unknown backend changes.

> This check was done with automatic power limiting [turned off by RyzenAdj](https://github.com/FlyGoat/RyzenAdj).

After doing that what wattage and frequencies are you seeing when running prompt processing? I'm hoping that you know what you're doing here and are not just casually cranking up the wattage and current limits as that can fry your chip.