@netrunnereve 2025-04-21
> I think something is jammed because there are only two graphics cores, too. I will treat this within local fork.

I don't understand what you mean here as from your `llama-bench` results you're definitely using all 8 of your GPU cores. Also if your GPU hypothetically had its core count limited you wouldn't be able to fix it by modifying llama.cpp code.

> I was monitoring watts, clock, load rate, etc. on the GPU-Z and the RyzenAdj, and it didn't look to me like there was any difference between the master and PR. I've attached the logs for part of the second.

Thanks, that's pretty helpful. I skimmed through the chart and I'm seeing a bit of power limiting during prompt processing with the chip hitting an average of 1700 MHz or so. It then jumps up to the full 1900 MHz for inference, and in both cases it's running slightly below the 25W limit. This is perfectly normal and I see this a lot on my own GPUs since the prompt processing stage is compute bound while inference is memory bound.

Considering how the prompt processing clocks are similar between master and my PR my guess is that your Windows driver is behaving differently from the Linux RADV driver that's used by @0cc4m and me.