@netrunnereve 2025-04-21
Speaking of limiting core count I retested this PR on my 470 with only 8 CUs enabled (2 per shader engine) and still got a 20% improvement in prompt processing speed. Yeah this is looking more and more like a driver thing.