@netrunnereve 2025-04-24
I also noticed that your driver is reporting that you have 32k shared memory on your Vega graphics which makes no sense. Anyways I've straight up disabled these changes for the AMD proprietary driver so we should be good to go.